8308960: Decouple internal Version and OperatingSystem classes
